web/requirements-dev.txt
pytest==9.0.3
pytest-flask==1.3.0
hypothesis==6.152.4
